app/requirements.txt
scrapy~=2.11.0
scrapy-user-agents
pandas~=2.1.1
pymysql~=1.1.0
schedule~=1.2.1
requests~=2.31.0
loguru~=0.7.2
pydantic~=2.4.2
uvicorn~=0.24.0.post1
fastapi~=0.104.1
python-dotenv~=1.0.0
starlette~=0.27.0
itemadapter~=0.8.0
mysql-connector-python~=8.1.0
numpy~=1.26.1
SQLAlchemy~=2.0.19
fasttext~=0.9.2
scipy
scikit-learn
sentry-sdk